+++ core/model/filters.go
@@ -5,6 +5,7 @@ type MembershipFilter struct {
 	ID         *string  `json:"id"`          // membership id
 	GroupIDs   []string `json:"group_ids"`   // list of group ids
 	UserID     *string  `json:"user_id"`     // core user id
+	UserIDs    []string `json:"user_ids"`    // core user ids
 	ExternalID *string  `json:"external_id"` // core user external id
 	NetID      *string  `json:"net_id"`      // core user net id
 	Name       *string  `json:"name"`        // member's name